Fixed role access
removed the BookServiceTests cuz my BookService is still a work in progress.

+++ src/main/java/nl/belastingdienst/library/book/BookService.java
@@ -57,7 +57,10 @@ public class BookService {
                 .collect(Collectors.toList());
     }
 
-    public void deleteBook(String ISBN13){
+    public void deleteBook(String ISBN13) throws Exception{
+        if (bookRepository.findById(ISBN13).isEmpty()) {
+            throw new Exception("Book doesn't exist already");
+        }
         bookRepository.deleteById(ISBN13);
     }
